Refactor alice and faber to use a common utils library
